Move exceptions to Cog\Contracts namespace
