+++ src/Authenticator/CookieAuthenticator.php
@@ -63,7 +63,7 @@ class CookieAuthenticator implements AuthenticatorContract
      * @param \Cog\Contracts\YouTrack\Rest\Client\Client $client
      * @return void
      *
-     * @throws \Cog\YouTrack\Rest\Authenticator\Exceptions\AuthenticationException
+     * @throws \Cog\Contracts\YouTrack\Rest\Authenticator\Exceptions\AuthenticationException
      */
     public function authenticate(ClientContract $client): void
     {
