+++ src/HttpClient/GuzzleHttpClient.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Cog\YouTrack\Rest\HttpClient;
 
 use Cog\Contracts\YouTrack\Rest\HttpClient\HttpClient as HttpClientContract;
-use Cog\YouTrack\Rest\HttpClient\Exceptions\HttpClientException;
+use Cog\Contracts\YouTrack\Rest\HttpClient\Exceptions\HttpClientException;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\BadResponseException;
 use GuzzleHttp\Exception\RequestException;
@@ -53,7 +53,7 @@ class GuzzleHttpClient implements HttpClientContract
      * @param array $options Additional Options
      * @return \Psr\Http\Message\ResponseInterface Raw response from the server
      *
-     * @throws \Cog\YouTrack\Rest\HttpClient\Exceptions\HttpClientException
+     * @throws \Cog\Contracts\YouTrack\Rest\HttpClient\Exceptions\HttpClientException
      */
     public function request(string $method, string $uri, array $options = []): ResponseInterface
     {
