+++ tests/Feature/Authenticator/TokenAuthorizerTest.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Cog\YouTrack\Rest\Tests\Feature\Authorizer;
 
-use Cog\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException;
+use Cog\Contracts\YouTrack\Rest\Authorizer\Exceptions\InvalidTokenException;
 use Cog\YouTrack\Rest\Authorizer\TokenAuthorizer;
 use Cog\YouTrack\Rest\Client\YouTrackClient;
 use Cog\YouTrack\Rest\HttpClient\GuzzleHttpClient;
